requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
idna==3.3
Pillow==9.0.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
